Pack accumulators even tighter

+++ contracts/test/accumulators/CurveLiquidityAccumulatorStub.sol
@@ -18,7 +18,7 @@ contract CurveLiquidityAccumulatorStub is CurveLiquidityAccumulator {
     function harnessFetchLiquidity(address token)
         public
         view
-        returns (uint256 tokenLiquidity, uint256 quoteTokenLiquidity)
+        returns (uint112 tokenLiquidity, uint112 quoteTokenLiquidity)
     {
         return super.fetchLiquidity(token);
     }
