+++ contracts/test/accumulators/UniswapV2LiquidityAccumulatorStub.sol
@@ -27,7 +27,7 @@ contract UniswapV2LiquidityAccumulatorStub is UniswapV2LiquidityAccumulator {
     function harnessFetchLiquidity(address token)
         public
         view
-        returns (uint256 tokenLiquidity, uint256 quoteTokenLiquidity)
+        returns (uint112 tokenLiquidity, uint112 quoteTokenLiquidity)
     {
         return super.fetchLiquidity(token);
     }
